feat: テーマ切り替え機能の追加とREADMEの更新
- ダッシュボードレイアウトにModeToggleコンポーネントを追加。
- ChatInputのダークモード背景色不一致問題を調査、修正を試行（修正はロールバック済み）。
- ChatHistoryコンポーネント内のReactMarkdown警告を修正。
- READMEを2025-11-20の開発履歴で更新:
  - 履歴削除機能とUI改善、ダークテーマ/トグルを追加。
  - 軽微なバグ修正を実施。
  - サイドバーをリファクタリングしコンポーネントを分割。
