+++ README.md
@@ -1,4 +1,4 @@
-# Castor - AI Pentest Assistant
+# CASTOR - AI Pentest Assistant
 
 ## プロジェクト概要
 - 本ツールは、サイバーセキュリティ学習支援を目的とした、安全性と適用性を重視したローカル完結型のAI支援アプリケーションです。
@@ -51,7 +51,6 @@
 - **UIライブラリ**: shadcn/ui
 - **スタイリング**: Tailwind CSS
 - **Markdownレンダリング**: react-markdown, remark-gfm
-- **シンタックスハイライト**: react-syntax-highlighter
 
 ## セットアップ手順
 
@@ -144,9 +143,12 @@ npm run dev
 - チャット開始フローを全面的にリファクタリングし、UI/UXを改善。
 - 
 **2025-11-16**:
-- 初期バージョン完成に向けた調整。
 - Google GenAI SDK準拠で設計したフェーズ別プロンプトを実装。
-- フロントエンドにMarkdownレンダリングとコードブロックのシンタックスハイライト機能を追加。
+- フロントエンドにMarkdownレンダリングを追加。
 - READMEを更新し、プロジェクトの目的、設計原則、セットアップ手順、技術スタックを明確化。
 - READMEに動作イメージのスクリーンショットを追加。
 - 
+**2025-11-20**:
+- 履歴削除機能とUI改善、ダークテーマ/トグルを追加。
+- 軽微なバグ修正を実施。
+- サイドバーをリファクタリングしコンポーネント分割。
