+++ frontend/package.json
@@ -24,12 +24,15 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.552.0",
+    "prism-react-renderer": "^2.4.1",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
     "react-markdown": "^10.1.0",
     "react-router-dom": "^7.9.5",
     "react-syntax-highlighter": "^16.1.0",
+    "rehype-pretty-code": "^0.14.1",
     "remark-gfm": "^4.0.1",
+    "shiki": "^3.15.0",
     "tailwind-merge": "^3.3.1",
     "tailwindcss": "^4.1.17",
     "zod": "^4.1.12"
@@ -39,6 +42,7 @@
     "@types/node": "^24.6.0",
     "@types/react": "^19.1.16",
     "@types/react-dom": "^19.1.9",
+    "@types/react-syntax-highlighter": "^15.5.13",
     "@vitejs/plugin-react": "^5.0.4",
     "eslint": "^9.36.0",
     "eslint-plugin-react-hooks": "^5.2.0",
