+++ frontend/src/components/chat-input.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { ArrowUpIcon } from "lucide-react";
 import { Textarea } from "@/components/ui/textarea";
 import { Button } from "@/components/ui/button";
-
+import type { Message } from '@/types/types';
 
 interface ChatInputProps {
   input: string;
@@ -13,7 +13,7 @@ interface ChatInputProps {
   handleSubmit: (e: React.FormEvent) => void;
   project: string;
   disabled?: boolean;
-  displayedHistory: { project: string; phase: string; sessionId: string; messages: any[] } | null;
+  displayedHistory: { project: string; phase: string; sessionId: string; messages: Message[] } | null;
   resumeDisplayedHistory: () => void;
   cancelHistoryDisplay: () => void;
 }
@@ -34,7 +34,7 @@ export const ChatInput: React.FC<ChatInputProps> = ({
 
   if (displayedHistory) {
     return (
-      <div className="flex flex-col border border-muted rounded-2xl shadow-lg bg-card p-2">
+      <div className="flex flex-col border border-border rounded-2xl shadow-lg bg-muted p-2">
         <div className="flex flex-col items-center justify-center p-4 gap-2">
           <p className="text-muted-foreground text-center">
             履歴を表示中: プロジェクト「{displayedHistory.project}」フェーズ「{displayedHistory.phase}」
@@ -54,18 +54,17 @@ export const ChatInput: React.FC<ChatInputProps> = ({
 
   return (
     <form onSubmit={handleSubmit}>
-
-      <div className="flex flex-col border border-muted rounded-2xl shadow-lg bg-card p-2">
+      <div className="flex flex-col border border-border rounded-2xl shadow-lg bg-card p-2">
         <Textarea
           value={input}
           onChange={(e) => setInput(e.target.value)}
           placeholder={placeholderText}
           disabled={disabled || isLoading}
-          className="bg-transparent text-card-foreground resize-none border-none focus-visible:ring-0 shadow-none min-h-[50px] p-2"
+          className="!bg-transparent !text-base placeholder:text-ms text-card-foreground resize-none border-none focus-visible:ring-0 shadow-none min-h-[50px] p-2"
         />
 
         <div className="flex items-center p-1">
-          <div className="flex-grow" />
+          <div className="flex-grow"/>
           <Button
             type="submit"
             size="icon"
