+++ frontend/src/global.css
@@ -42,7 +42,7 @@
 .dark {
   --background:  oklch(0.145 0 0);
   --foreground:  oklch(0.985 0 0);
-  --card:  oklch(0.205 0 0);
+  --card:  oklch(1 0 0 / 7%);
   --card-foreground:  oklch(0.985 0 0);
   --popover:  oklch(0.205 0 0);
   --popover-foreground:  oklch(0.985 0 0);
@@ -117,7 +117,7 @@
     @apply border-border outline-ring/50;
   }
   body {
-    @apply bg-background text-foreground;
+    @apply bg-background text-foreground font-sans;
   }
 }
