+++ frontend/tailwind.config.js
@@ -85,6 +85,9 @@ module.exports = {
         "accordion-down": "accordion-down 0.2s ease-out",
         "accordion-up": "accordion-up 0.2s ease-out",
       },
+      fontFamily: {
+        'sans': ['BIZ UDGothic', 'sans-serif'],
+      },
     },
   },
   plugins: [require("tailwindcss-animate")],
